Shells/selenium/src/requirements.txt
cloudshell-shell-core>=3.1,<4.0
cloudshell-core
cloudshell-automation-api
selenium==3.3.1
requests
pygithub